refactor(#130): clients 디렉토리 재생성
