+++ apps/web/src/clients/index.ts
@@ -1,0 +1,4 @@
+export { default as Home } from './home/home';
+export { default as SignIn } from './signin/signin';
+export { default as SignUp } from './signup/signup';
+export { default as NotFound } from './notfound/notfound';
