+++ apps/web/src/components/notfound/index.ts
@@ -1,3 +1,3 @@
-export { default as NotFoundContent } from '@/clients/NotFoundContent/NotFoundContent';
+export { default as NotFoundContent } from '@/clients/notfound/notfound';
 export { default as SlotMachineDigits } from './SlotMachineDigits/SlotMachineDigits';
 export { default as WinMessage } from './WinMessage/WinMessage';
