+++ apps/web/src/components/signin/index.ts
@@ -1,3 +1,3 @@
-export { default as SignInContent } from '@/clients/SignInContent/SignInContent';
+export { default as SignInContent } from '@/clients/signin/signin';
 export { default as ContinueWithGoogle } from './ContinueWithGoogle/ContinueWithGoogle';
 export { default as SecondaryAction } from './SecondaryAction/SecondaryAction';
